+++ src/controllers/RoleController.ts
@@ -4,7 +4,7 @@ import { Role } from "../entity/Role.js";
 class RoleController {
   static getRoles = async (request: any, response: any) => {
     try {
-      const results = await pool.query('SELECT * FROM public.role');
+      const results = await pool.query("SELECT * FROM public.role");
       const roles: Role[] = results.rows.map((row) => {
         return new Role(
           row.id.toString(),
@@ -21,9 +21,10 @@ class RoleController {
   };
   static getRoleById = async (roleId: string) => {
     try {
-      const result = await pool.query('SELECT * FROM public.role WHERE id = $1', [
-        roleId,
-      ]);
+      const result = await pool.query(
+        "SELECT * FROM public.role WHERE id = $1",
+        [roleId]
+      );
       if (result.rows.length === 0) {
         throw new Error("Role not found");
       }
@@ -45,12 +46,15 @@ class RoleController {
     const { id, title } = request.body;
 
     try {
-      const result = await pool.query('SELECT * FROM public.role WHERE id = $1', [id]);
+      const result = await pool.query(
+        "SELECT * FROM public.role WHERE id = $1",
+        [id]
+      );
 
       if (result.rows.length > 0) {
         const role = result.rows[0];
         await pool.query(
-          'UPDATE public.role SET title = $1, updated_at = NOW() WHERE id = $2',
+          "UPDATE public.role SET title = $1, updated_at = NOW() WHERE id = $2",
           [title, id]
         );
 
@@ -64,7 +68,7 @@ class RoleController {
         return response.status(200).json(updatedRole);
       } else {
         const newRole = await pool.query(
-          'INSERT INTO public.role (title, created_at, updated_at) VALUES ($1, NOW(), NOW()) RETURNING *',
+          "INSERT INTO public.role (title, created_at, updated_at) VALUES ($1, NOW(), NOW()) RETURNING *",
           [title]
         );
 
@@ -86,13 +90,27 @@ class RoleController {
     }
   };
 
-    static deleteRoleById = async (roleId: string) => {
-        try {
-            await pool.query('DELETE FROM public.role WHERE id=$1', [roleId]);
-        } catch (error) {
-            throw error;
-        }
+  static deleteRoleById = async (roleId: string) => {
+    try {
+      const result = await pool.query(
+        "SELECT * FROM public.role WHERE id = $1",
+        [roleId]
+      );
+
+      if (result.rows.length > 0) {
+        await pool.query("DELETE FROM public.role WHERE id = $1", [roleId]);
+        return {
+          success: true,
+          message: "Role deleted successfully",
+          data: result.rows[0],
+        };
+      } else {
+        throw new Error("role not found");
+      }
+    } catch (error) {
+      throw new Error("An error occurred while deleting the role");
     }
+  };
 }
 
 export { RoleController };
